Move action handlers to hackbot-api
