+++ services/hackbot-api/app/action_handlers/__init__.py
@@ -1,15 +1,15 @@
 """Apply-side handlers for recorded actions."""
 
-from hackbot_runtime.actions.handlers.base import (
+from app.action_handlers.base import (
     ActionHandler,
     ActionResult,
     ApplyContext,
 )
-from hackbot_runtime.actions.handlers.bugzilla_handler import (
+from app.action_handlers.bugzilla_handler import (
     merge_resolved,
     plan_coalesced_groups,
 )
-from hackbot_runtime.actions.handlers.registry import HANDLERS, get_handler
+from app.action_handlers.registry import HANDLERS, get_handler
 
 __all__ = [
     "ActionHandler",
