+++ services/hackbot-api/app/action_handlers/bugzilla_handler.py
@@ -17,7 +17,7 @@ from typing import Any
 
 import requests
 
-from hackbot_runtime.actions.handlers.base import ActionResult, ApplyContext
+from app.action_handlers.base import ActionResult, ApplyContext
 
 log = logging.getLogger(__name__)
