+++ services/hackbot-api/app/action_handlers/phabricator_handler.py
@@ -27,7 +27,7 @@ from typing import Any
 from async_lru import alru_cache
 from phabricator_client import PhabricatorClient
 
-from hackbot_runtime.actions.handlers.base import ActionResult, ApplyContext
+from app.action_handlers.base import ActionResult, ApplyContext
 
 log = logging.getLogger(__name__)
